requirements.txt
# Core dependencies
streamlit>=1.29.0
google-genai
google-generativeai>=0.3.2
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
